+++ drizzle/0007_strong_true_believers.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `tasks` ADD `notify` integer DEFAULT true NOT NULL;--> statement-breakpoint
+ALTER TABLE `treatments` ADD `notify` integer DEFAULT true NOT NULL;